V2

+++ interface.py
@@ -35,9 +35,27 @@ def SaveFile():
             f.write(contingut)
         print(f"Arxiu desat a: {filepath}")
 
-    fitxer_temporal = selected_file
-    if os.path.exists(fitxer_temporal):
-        os.remove(fitxer_temporal)
+    if selected_file!="ElsMeusNodesSegments.txt":
+        fitxer_temporal = selected_file
+        if os.path.exists(fitxer_temporal):
+            os.remove(fitxer_temporal)
+
+def CarregarFicher():
+    global selected_file
+    ruta_fitxer = filedialog.askopenfilename(
+        title="Selecciona un fitxer .txt",
+        filetypes=[("Fitxers de text", "*.txt")]
+    )
+    if ruta_fitxer:
+        print("Has seleccionat:", ruta_fitxer)
+        with open(ruta_fitxer, "r", encoding="utf-8") as fitxer:
+            contingut = fitxer.read()
+            print("Contingut del fitxer:")
+            print(contingut)
+
+        selected_file="document_nou.txt"
+        with open(selected_file, "w") as f:
+            f.writelines(contingut)
 
 def showtext():
     messagebox.showinfo("Text introduït: ", entry.get())
@@ -50,15 +68,6 @@ def showgrafoexemple():
 
 def showgrafoexemple_vaeïns():
     A = CreateGraph_1()
-    node_origin_for_grafos=entry.get()
-    PlotNode(A, node_origin_for_grafos)
-
-def showgrafoinventat():
-    A = CreateGraph_2()
-    Plot(A)
-
-def showgrafoinventat_vaeïns():
-    A = CreateGraph_2()
     node_origin_for_grafos=entry.get()
     PlotNode(A, node_origin_for_grafos)
 
@@ -126,8 +135,6 @@ def DeleteSegmentToTheFile():
 #def SaveGraph():
 
 
-
-
 #li donem  mida i títol a la finestra i creem les files i columnes necessàries //estructura de la finestra
 fin=tk.Tk()
 fin.geometry("1000x600")
@@ -149,28 +156,20 @@ graficsexemple.rowconfigure(1, weight=1)
 graficsexemple.columnconfigure(0, weight=1)
 
 button1=tk.Button(graficsexemple, text="Mapa grafo", command=showgrafoexemple)
-button1.grid(row=0, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
+button1.grid(row=0, column=0, padx=4, pady=4, sticky=tk.N + tk.E + tk.W + tk.S)
 
 button2=tk.Button(graficsexemple, text="Grafo amb veïns", command=showgrafoexemple_vaeïns)
-button2.grid(row=1, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
+button2.grid(row=1, column=0, padx=4, pady=4, sticky=tk.N + tk.E + tk.W + tk.S)
 
-#Configuració espai grafics inventat
-graficinventat=tk.LabelFrame(fin,text="El nostre grafo inventat:")
-graficinventat.grid(row=1,column=0,pady=10,padx=10,sticky=tk.N + tk.E + tk.W + tk.S)
+button3=tk.Button(graficsexemple, text="Reachablity Map", command=showgrafoexemple)
+button3.grid(row=2, column=0, padx=4, pady=4, sticky=tk.N + tk.E + tk.W + tk.S)
 
-graficinventat.rowconfigure(0, weight=1)
-graficinventat.rowconfigure(1, weight=1)
-graficinventat.columnconfigure(0, weight=1)
-
-button1=tk.Button(graficinventat, text="Mapa grafo", command=showgrafoinventat)
-button1.grid(row=0, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
-
-button2=tk.Button(graficinventat, text="Grafo amb veïns", command=showgrafoinventat_vaeïns)
-button2.grid(row=1, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
+button4=tk.Button(graficsexemple, text="Shortest Path", command=showgrafoexemple_vaeïns)
+button4.grid(row=3, column=0, padx=4, pady=4, sticky=tk.N + tk.E + tk.W + tk.S)
 
 #Configuració espai grafics del arxiu text
 grafictext=tk.LabelFrame(fin,text="Grafo llegint un document:")
-grafictext.grid(row=2,column=0,pady=10,padx=10,sticky=tk.N + tk.E + tk.W + tk.S)
+grafictext.grid(row=1,column=0,pady=10,padx=10,sticky=tk.N + tk.E + tk.W + tk.S)
 
 grafictext.rowconfigure(0, weight=1)
 grafictext.rowconfigure(1, weight=1)
@@ -182,8 +181,6 @@ button1.grid(row=0, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
 button2=tk.Button(grafictext, text="Grafo amb veïns", command=showgrafofromfiles_veins)
 button2.grid(row=1, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
 
-#button3=tk.Button(espaigrafics, text="Grafic arxiu")
-#button3.grid(row=2, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
 
 #Configuració espai pels input buscar node
 inputframe=tk.LabelFrame(fin,text="Tria del node d'origen:")
@@ -262,6 +259,7 @@ guardat.grid(row=2,column=2,pady=10,padx=10,sticky=tk.N + tk.E + tk.W + tk.S)
 guardat.rowconfigure(0, weight=1)
 guardat.rowconfigure(1, weight=1)
 guardat.rowconfigure(2, weight=1)
+guardat.rowconfigure(3, weight=1)
 guardat.columnconfigure(0, weight=1)
 
 entry_file=tk.Entry(guardat)
@@ -272,5 +270,11 @@ button.grid(row=1, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
 
 button=tk.Button(guardat, text="Guardar arxiu seleccionat", command=SaveFile)
 button.grid(row=2, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
+
+button_charge=tk.Button(guardat, text="Input", command=SelectedFile)
+button_charge.grid(row=1, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
+
+button_charge=tk.Button(guardat, text="Carregar fichero", command=CarregarFicher)
+button_charge.grid(row=3, column=0, padx=5, pady=5, sticky=tk.N + tk.E + tk.W + tk.S)
 
 fin.mainloop()
